Show "<unknown>" rather than "None" in top entry contributors (#123)
Some members left the Replit Discord, preventing the website from showing their username correctly in the "top entry contributors" and instead showing `None`. This fixes it so it says <unknown> instead of None, making it look less like a bug. I also fixed a small issue in the css which was causing an underline to include more than it should.